pyproject.toml
[project]
name = "langchain-20250313"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "meijunjie@kzl.com.cn" }
]
requires-python = ">=3.10"
dependencies = [
    "dashscope>=1.22.2",
    "fastapi>=0.115.11",
    "langchain>=0.3.20",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.19",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.8",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.34.0",
    "torch==2.5.1+cu121",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "transformers>=4.49.0",
    "modelscope>=1.23.2",
    "setuptools>=76.0.0",
    "aiomysql>=0.2.0",
    "datasets>=3.4.0",
]

[project.scripts]
server = "langchain_20250313:server"
cuda = "langchain_20250313:cuda"
part2 = "langchain_20250313:part2"
part4 = "langchain_20250313:part4"
part6 = "langchain_20250313:part6"
part9 = "langchain_20250313:part9"
part13 = "langchain_20250313:part13"
part13_2 = "langchain_20250313:part13_2"
part17 = "langchain_20250313:part17"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[[tool.uv.index]]
name="torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name="torchaudio"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name="torchvision"
url = "https://download.pytorch.org/whl/cu121"
explicit = true


[tool.uv.sources]
torch = [
  { index = "torch_cu121"}
]
torchvision = [
  { index = "torch_cu121"}
]

torchaudio = [
  { index = "torch_cu121"}
]
